requirements.txt
# 基础依赖
tkinter  # 通常内置于Python
requests>=2.25.1
Pillow>=8.0.0
 
# Markdown支持
markdown>=3.3.4

# WebView支持
pywebview>=3.6.3 